Broke out and grouped "leasing-service"-endpoints based on root resource
