+++ packages/frontend/src/pages/ParkingSpace/hooks/useRemoveApplicantFromListing.ts
@@ -9,13 +9,16 @@ export const useRemoveApplicantFromListing = () => {
   const queryClient = useQueryClient()
   return useMutation<unknown, AxiosError, Params>({
     mutationFn: (params: Params) =>
-      axios.delete<unknown>(`${backendUrl}/applicant/${params.applicantId}`, {
-        headers: {
-          Accept: 'application/json',
-          'Access-Control-Allow-Credentials': true,
-        },
-        withCredentials: true,
-      }),
+      axios.delete<unknown>(
+        `${backendUrl}/listings/applicant/${params.applicantId}`,
+        {
+          headers: {
+            Accept: 'application/json',
+            'Access-Control-Allow-Credentials': true,
+          },
+          withCredentials: true,
+        }
+      ),
     onSuccess: (_, params) =>
       queryClient.invalidateQueries({
         queryKey: ['parkingSpaceListing', params.listingId],
